feat(domain): 프로모션 클래스를 추상화 하여 재사용 가능 하도록 업그레이드
